+++ scraper.js
@@ -1,14 +1,28 @@
 const async = require("async-q"),
   config = require("./config"),
-  eztv = require("./providers/eztv")("EZTV"),
+  eztv = require("./providers/show/eztv")("EZTV"),
+  katMovie = require("./providers/movie/kat"),
+  katShow = require("./providers/show/kat"),
   util = require("./util");
 
-/* Start scraping from KAT. */
-const scrapeKAT = () => {
-  return async.eachSeries(config.providers, (provider) => {
+/* Start movie scraping from KAT. */
+const scrapeKatMovies = () => {
+  return async.eachSeries(config.movieProviders, (provider) => {
     util.setStatus("Scraping " + provider.name);
-    const kat = require("./providers/kat")(provider.name);
-    return util.spawn(kat.search(provider)).then((response) => {
+    const katProvider = katMovie(provider.name);
+    return util.spawn(katProvider.search(provider)).then((response) => {
+      console.log(provider.name + ": Done.");
+      return response;
+    });
+  });
+};
+
+/* Start show scraping from KAT. */
+const scrapeKATShows = () => {
+  return async.eachSeries(config.showProviders, (provider) => {
+    util.setStatus("Scraping " + provider.name);
+    const katProvider = katShow(provider.name);
+    return util.spawn(katProvider.search(provider)).then((response) => {
       console.log(provider.name + ": Done.");
       return response;
     });
@@ -31,10 +45,11 @@ module.exports = {
     util.resetTemp();
     util.setlastUpdate();
 
-    async.eachSeries([scrapeEZTV, scrapeKAT], (scraper) => {
+    async.eachSeries([scrapeEZTV, scrapeKATShows, scrapeKatMovies], (scraper) => {
       return scraper();
     }).catch((err) => {
       util.onError("Error while scraping: " + err);
+      console.log(err);
       return err;
     }).done();
   }